requirements.txt
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
datasetops==0.0.6
decorator==5.1.1
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.18
idna==3.4
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
kaggle==1.5.12
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
office31==0.1.3
pandas==1.1.5
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
sentry-sdk==1.12.0
setproctitle==1.2.3
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.13
wandb==0.13.7
wcwidth==0.2.5
zipp==3.6.0

